# server/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/logs": {
      "get": {
        "tags": [
          "logs-controller"
        ],
        "operationId": "getLogs",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "description": "Filter by log level (DEBUG, INFO, WARN, ERROR) - case insensitive",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date-time in ISO-8601 format (e.g., '2024-02-14T10:30:00' or '2024-02-14T10:30:00Z')",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "2024-02-14T10:30:00"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date-time in ISO-8601 format (e.g., '2024-02-14T15:30:00' or '2024-02-14T15:30:00Z')",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "2024-02-14T15:30:00"
          },
          {
            "name": "logger",
            "in": "query",
            "description": "Filter by logger name (substring match, case insensitive)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "in": "query",
            "description": "Filter by MDC requestId (exact match, case insensitive) to trace all logs of one request",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LogsResponseWrapper"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/library/movies": {
      "get": {
        "tags": [
          "library-movies-controller"
        ],
        "operationId": "getMovies",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "filters",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MoviesFiltersDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MoviesResponseWrapper"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/library/movies/watched/unrated": {
      "get": {
        "tags": [
          "library-movies-controller"
        ],
        "operationId": "getRecentlyWatchedUnratedMovies",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "filters",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MoviesFiltersDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MoviesResponseWrapper"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/library/movies/broken": {
      "get": {
        "tags": [
          "library-movies-controller"
        ],
        "operationId": "getBrokenMovies",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BrokenMediaItemResponseWrapper"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaginationMetadata": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "isFirst": {
            "type": "boolean"
          },
          "isLast": {
            "type": "boolean"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
...
      "BrokenMediaItemResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "jellyfinId": {
            "type": "string"
          },
          "TMDbId": {
            "type": "string"
          },
          "IMDbId": {
            "type": "string"
          },
          "productionYear": {
            "type": "integer",
            "format": "int32"
          },
          "missingFields": {
            "type": "string"
          },
          "resolvedMediaItemId": {
            "type": "string",
            "format": "uuid"
          },
          "resolved": {
            "type": "boolean"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CastMemberResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "person": {
            "$ref": "#/components/schemas/PersonResponseDTO"
          },
          "character": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CrewMemberResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "person": {
            "$ref": "#/components/schemas/PersonResponseDTO"
          },
          "job": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "profilePathUrl": {
            "type": "string"
          }
        }
      },
...
      "GenreResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "MediaItemMetadataResponseDTO": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "posterImageUrl": {
            "type": "string"
          },
          "backdropImageUrl": {
            "type": "string"
          },
          "imdbRating": {
            "type": "number",
            "format": "float"
          },
          "rottenTomatoesRating": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "MovieResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "jellyfinId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "TMDbId": {
            "type": "string"
          },
          "IMDbId": {
            "type": "string"
          },
          "productionYear": {
            "type": "integer",
            "format": "int32"
          },
          "runtimeMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "metadata": {
            "$ref": "#/components/schemas/MediaItemMetadataResponseDTO"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreResponseDTO"
            },
            "uniqueItems": true
          },
          "cast": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CastMemberResponseDTO"
            }
          },
          "crew": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CrewMemberResponseDTO"
            }
          },
          "myRating": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "PersonResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "TMDbId": {
            "type": "string"
          },
          "profilePathUrl": {
            "type": "string"
          }
        }
      },
...
      "LogEvent": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "level": {
            "type": "string"
          },
          "logger": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "mdc": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "serviceName": {
            "type": "string"
          },
          "thread": {
            "type": "string"
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "stackTrace": {
            "type": "string"
          }
        }
      },
      "LogsResponseWrapper": {
        "type": "object",
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogEvent"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetadata"
          }
        }
      },
...
      "BrokenMediaItemResponseWrapper": {
        "type": "object",
        "properties": {
          "movies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrokenMediaItemResponseDTO"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetadata"
          }
        }
      },
...
      "MoviesFiltersDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "releasedAfter": {
            "type": "integer",
            "format": "int32"
          },
          "releasedBefore": {
            "type": "integer",
            "format": "int32"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "posterSize": {
            "type": "string"
          },
          "backdropSize": {
            "type": "string"
          },
          "profileSize": {
            "type": "string"
          },
          "profileId": {
            "type": "string",
            "format": "uuid"
          },
          "ratingMin": {
            "type": "number",
            "format": "float"
          },
          "ratingMax": {
            "type": "number",
            "format": "float"
          },
          "personId": {
            "type": "string",
            "format": "uuid"
          },
          "unrated": {
            "type": "boolean"
          }
        },
        "required": [
          "title"
        ]
      },
      "MoviesResponseWrapper": {
        "type": "object",
        "properties": {
          "movies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovieResponseDTO"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetadata"
          }
        }
      },
...
    }
  }
}